feat(board): sync board with shared config state
